requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distlib==0.3.8
distro==1.9.0
filelock==3.13.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2023.12.0
googlemaps==4.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.48
MarkupSafe==2.1.3
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.19.0
orjson==3.10.1
packaging==23.2
platformdirs==4.2.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
torch==2.1.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.2
yarl==1.9.4
